+++ src/main/java/com/edu/nju/alley/dao/support/PostDSS.java
@@ -32,14 +32,14 @@ public final class PostDSS {
 
     public static final SqlColumn<Integer> anchorId = post.anchorId;
 
-    public static final SqlColumn<Integer> longitude = post.longitude;
+    public static final SqlColumn<Float> longitude = post.longitude;
 
-    public static final SqlColumn<Integer> latitude = post.latitude;
+    public static final SqlColumn<Float> latitude = post.latitude;
 
     public static final SqlColumn<String> pictureUrl = post.pictureUrl;
 
     public static final class Post extends SqlTable {
-        
+
         public final SqlColumn<Integer> id = column("id", JDBCType.INTEGER);
 
         public final SqlColumn<Integer> authId = column("auth_id", JDBCType.INTEGER);
@@ -62,9 +62,9 @@ public final class PostDSS {
 
         public final SqlColumn<Integer> anchorId = column("anchor_id", JDBCType.INTEGER);
 
-        public final SqlColumn<Integer> longitude = column("longitude", JDBCType.INTEGER);
+        public final SqlColumn<Float> longitude = column("longitude", JDBCType.FLOAT);
 
-        public final SqlColumn<Integer> latitude = column("latitude", JDBCType.INTEGER);
+        public final SqlColumn<Float> latitude = column("latitude", JDBCType.FLOAT);
 
         public final SqlColumn<String> pictureUrl = column("picture_url", JDBCType.VARCHAR);
