+++ src/main/java/com/edu/nju/alley/dto/PostDTO.java
@@ -15,9 +15,9 @@ public class PostDTO {
 
     private Integer anchorId;
 
-    private Integer longitude;
+    private Float longitude;
 
-    private Integer latitude;
+    private Float latitude;
 
     private PostAuthDTO auth;
