+++ src/main/java/com/edu/nju/alley/po/Post.java
@@ -33,9 +33,9 @@ public class Post {
 
     private Integer anchorId;
 
-    private Integer longitude;
+    private Float longitude;
 
-    private Integer latitude;
+    private Float latitude;
 
     private String pictureUrl;
